Move reduce tests out of ctest_array and into ctest_array_reduce.
